+++ src/Sylius/Bundle/AdminBundle/CommandHandler/CreateAdminUserHandler.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AdminBundle\MessageHandler;
+namespace Sylius\Bundle\AdminBundle\CommandHandler;
 
 use Sylius\Bundle\AdminBundle\Exception\CreateAdminUserFailedException;
-use Sylius\Bundle\AdminBundle\Message\CreateAdminUser;
+use Sylius\Bundle\AdminBundle\Command\CreateAdminUser;
 use Sylius\Component\Core\Model\AdminUserInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
