+++ src/Sylius/Bundle/AdminBundle/Resources/config/services.xml
@@ -44,7 +44,7 @@
             <tag name="console.command" />
         </service>
 
-        <service id="Sylius\Bundle\AdminBundle\MessageHandler\CreateAdminUserHandler">
+        <service id="Sylius\Bundle\AdminBundle\CommandHandler\CreateAdminUserHandler">
             <argument type="service" id="sylius.repository.admin_user" />
             <argument type="service" id="sylius.factory.admin_user" />
             <argument type="service" id="sylius.canonicalizer" />
