+++ src/Sylius/Bundle/AdminBundle/Tests/CommandHandler/CreateAdminUserHandlerTest.php
@@ -11,13 +11,13 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AdminBundle\Tests\MessageHandler\Admin;
+namespace Sylius\Bundle\AdminBundle\Tests\CommandHandler;
 
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Sylius\Bundle\AdminBundle\Command\CreateAdminUser;
+use Sylius\Bundle\AdminBundle\CommandHandler\CreateAdminUserHandler;
 use Sylius\Bundle\AdminBundle\Exception\CreateAdminUserFailedException;
-use Sylius\Bundle\AdminBundle\Message\CreateAdminUser;
-use Sylius\Bundle\AdminBundle\MessageHandler\CreateAdminUserHandler;
 use Sylius\Component\Core\Model\AdminUserInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
