+++ src/Sylius/Bundle/AdminBundle/Tests/Console/Command/CreateAdminUserCommandTest.php
@@ -18,7 +18,7 @@ use PHPUnit\Framework\TestCase;
 use Sylius\Bundle\AdminBundle\Console\Command\CreateAdminUserCommand;
 use Sylius\Bundle\AdminBundle\Console\Command\Factory\QuestionFactoryInterface;
 use Sylius\Bundle\AdminBundle\Exception\CreateAdminUserFailedException;
-use Sylius\Bundle\AdminBundle\Message\CreateAdminUser;
+use Sylius\Bundle\AdminBundle\Command\CreateAdminUser;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Tester\CommandTester;
